Robustify some stuff.

+++ FindFiles.cs
@@ -10,33 +10,38 @@ namespace WallpaperManager
         public static void DoFindFiles()
         {
             var dbFile = GetDbFile();
+	        if (dbFile == "")
+		        return;
+
+	        var targetDirectory = Environment.GetFolderPath(Environment.SpecialFolder.MyPictures) + "\\ToolOutput";
+	        if (Directory.Exists(targetDirectory))
+		        Directory.Delete(targetDirectory, true);
+
             var allLines = File.ReadAllLines(dbFile);
-	        var query = allLines
-		        .Skip(1)
-		        .Select(WallpaperData.Parse);
-
             var basePath = Path.GetDirectoryName(dbFile) + "\\";
-            var files = query.Where(x => 
-	            !x.funny &&
-				!x.edgy &&
-				!x.gaming &&
-				!x.christmas &&
-				!x.dark &&
-	            x.environment.Contains(WallpaperData.Environment.urban));
 
-            var targetDirectory = Environment.GetFolderPath(Environment.SpecialFolder.MyPictures) + "\\ToolOutput";
-            if (Directory.Exists(targetDirectory))
-                Directory.Delete(targetDirectory, true);
+	        var targetFiles = allLines
+		        .Skip(1)
+		        .Select(WallpaperData.Parse)
+		        .Where(x =>
+			        !x.funny &&
+			        !x.edgy &&
+			        !x.gaming &&
+			        !x.christmas &&
+			        !x.dark &&
+			        x.HasProperty(WallpaperData.Environment.urban)).ToArray();
+
+	        if (targetFiles.Length == 0)
+		        return;
 
             Directory.CreateDirectory(targetDirectory);
 
-            int counter = 1;
-            foreach (var f in files)
+            var counter = 0;
+            foreach (var f in targetFiles)
             {
                 var sourcefile = basePath + f.filename;
-                var targetFile = targetDirectory + $"\\{counter}." + Path.GetFileName(sourcefile).Split('.')[1];
+                var targetFile = targetDirectory + $"\\{++counter}." + Path.GetFileName(sourcefile).Split('.')[1];
                 File.Copy(sourcefile, targetFile);
-                counter++;
             }
         }
